jenkins job tar log complete

+++ ironic-xcat-test/console_log.sh
@@ -10,13 +10,18 @@ function job_log {
     if [ -z "$ZUUL_PIPELINE" ]; then
         export ZUUL_PIPELINE="check"
     fi
-    export ZUUL_LOG_DIR="$ZUUL_PIPELINE_$ZUUL_CHANGE"
+    export ZUUL_LOG_DIR=${JOB_NAME}_${BUILD_NUMBER}
     mkdir $ZUUL_LOG_DIR
-    scp -i /opt/ci_tmp/id_rsa -rf jenkins@9.114.34.161:/opt/stack/logs.tar.gz $ZUUL_LOG_DIR
+    scp -i /opt/ci_tmp/id_rsa jenkins@9.114.34.161:/opt/stack/logs.tar.gz $ZUUL_LOG_DIR
 }
 
 job_log
 cd /tmp/ironic-xcat-test/logs/$DATE_DIR/$ZUUL_LOG_DIR
-rm consoleText*
+tar xvfz logs.tar.gz
+$console_result=`ls consoleTest*`
+if [[ -n "$console_result" ]]; then
+    rm consoleText*
+fi
+
 wget http://9.114.34.160:8080/job/$JOB_NAME/$BUILD_NUMBER/consoleText
 mv consoleText "console_$JOB_NAME_$BUILD_NUMBER.log"
